Rename classes as modules
